Changed DataModels, Recreated Database, Repositories fix, Services fix

+++ CryptoSimulator/CryptoSimulator.DataModels/Models/Wallet.cs
@@ -3,7 +3,7 @@ namespace CryptoSimulator.DataModels.Models
     public class Wallet
     {
         public int Id { get; set; }
-        public double MaxCoins { get; set; }
+        public int MaxCoins { get; set; }
         public double Cash { get; set; }
         public int UserId { get; set; }
         public User User { get; set; }
